auth-backend/requirements.txt
flask==3.0.0
requests>=2.28.0
PyJWT==2.8.0
werkzeug==3.0.1
flask_cors
pymongo